+++ shoppingApp/src/components/productCard/BrandCard.tsx
@@ -1,5 +1,5 @@
 import ProductImage from './ProductImage';
-import { imageWidth } from './ImageWH';
+import { imageWidth } from '../../variable/ImageWH';
 import CardProps from '../../types/CardProps';
 import { useState } from 'react';
 import Modal from '../modal/Modal';
